+++ src/main/java/com/mediscreennote/repository/NoteRepository.java
@@ -9,5 +9,5 @@ import com.mediscreennote.model.Note;
 
 @Repository
 public interface NoteRepository extends MongoRepository<Note, String> {
-	public List<Note> findByPatientId(Long patientId);
+	public List<Note> findByPatientId(int patientId);
 }
